feat: move TokenAccounts into features
